Decompose the tuple-match desugars and split the deep specializer
