+++ crates/almide-mir/src/lower/desugar_match_b.rs
@@ -232,6 +232,42 @@ fn introduces_binder(e: &IrExpr) -> bool {
 /// `_` or a tuple of Wildcard / `none` / fieldless-ctor components (no binds —
 /// exhaustiveness is the frontend's guarantee, the same last-arm-else discipline
 /// as every match lowering).
+/// Hoist each non-Var tuple component ONCE into a temp; a Var component
+/// is used direct. Returns the hoist binds and the per-component refs.
+fn hoist_tuple_components(
+    elements: &[IrExpr],
+    next: &mut u32,
+    span: &Option<almide_ir::Span>,
+) -> (Vec<IrStmt>, Vec<IrExpr>) {
+    // Hoist each non-Var component ONCE into a temp (a Var component is used direct).
+    let mut stmts: Vec<IrStmt> = Vec::new();
+    let mut refs: Vec<IrExpr> = Vec::new();
+    for c in elements {
+        if matches!(c.kind, IrExprKind::Var { .. }) {
+            refs.push(c.clone());
+        } else {
+            let t = VarId(*next);
+            *next += 1;
+            stmts.push(IrStmt {
+                kind: IrStmtKind::Bind {
+                    var: t,
+                    ty: c.ty.clone(),
+                    value: c.clone(),
+                    mutability: almide_ir::Mutability::Let,
+                },
+                span: span.clone(),
+            });
+            refs.push(IrExpr {
+                kind: IrExprKind::Var { id: t },
+                ty: c.ty.clone(),
+                span: span.clone(),
+                def_id: None,
+            });
+        }
+    }
+    (stmts, refs)
+}
+
 pub fn desugar_tuple_variant_match(body: &IrExpr) -> Option<IrExpr> {
     use almide_ir::visit_mut::{walk_expr_mut, IrMutVisitor};
     use almide_ir::IrPattern;
@@ -245,74 +281,109 @@ pub fn desugar_tuple_variant_match(body: &IrExpr) -> Option<IrExpr> {
                 | IrPattern::Constructor { .. }
         ) || matches!(p, IrPattern::List { elements } if elements.is_empty())
     }
-    struct V {
-        next: u32,
-        changed: bool,
+
+    /// The nesting context of [`nest_conditional_columns`] — the match being
+    /// rewritten and the per-column material, which only travel together.
+    struct NestCtx<'a> {
+        e: &'a IrExpr,
+        span: Option<almide_ir::Span>,
+        pats: &'a [IrPattern],
+        refs: &'a [IrExpr],
+        els: &'a IrExpr,
+        conditional: fn(&IrPattern) -> bool,
     }
-    impl IrMutVisitor for V {
-        fn visit_expr_mut(&mut self, e: &mut IrExpr) {
-            walk_expr_mut(self, e);
-            let IrExprKind::Match { subject, arms } = &e.kind else { return };
-            let IrExprKind::Tuple { elements } = &subject.kind else { return };
+
+    /// Nest the conditional components right-to-left (leftmost test
+    /// outermost) around `inner`.
+    fn nest_conditional_columns(cx: NestCtx<'_>, mut inner: IrExpr) -> IrExpr {
+        // Nest the conditional components right-to-left (leftmost test outermost).
+        for (i, p) in cx.pats.iter().enumerate().rev() {
+            if !(cx.conditional)(p) {
+                continue;
+            }
+            inner = IrExpr {
+                kind: IrExprKind::Match {
+                    subject: Box::new(cx.refs[i].clone()),
+                    arms: vec![
+                        almide_ir::IrMatchArm {
+                            pattern: p.clone(),
+                            guard: Option::None,
+                            body: inner,
+                        },
+                        almide_ir::IrMatchArm {
+                            pattern: IrPattern::Wildcard,
+                            guard: Option::None,
+                            body: cx.els.clone(),
+                        },
+                    ],
+                },
+                ty: cx.e.ty.clone(),
+                span: cx.span.clone(),
+                def_id: cx.e.def_id,
+            };
+        }
+        inner
+    }
+
+    /// Count the conditional columns of arm 1 and gate arm 2 as a valid
+    /// default (bare `_`, or an all-trivial tuple of the same width). `None`
+    /// = the shape declines.
+    fn refinement_column_census(
+        pats: &[IrPattern],
+        default_pat: &IrPattern,
+        conditional: fn(&IrPattern) -> bool,
+    ) -> Option<usize> {
+        let mut cond_n = 0usize;
+        for p in pats {
+            if conditional(p) {
+                cond_n += 1;
+            } else if !matches!(p, IrPattern::Wildcard | IrPattern::Bind { .. }) {
+                return None;
+            }
+        }
+        if cond_n == 0 {
+            return None;
+        }
+        match default_pat {
+            IrPattern::Wildcard => {}
+            IrPattern::Tuple { elements: p2 }
+                if p2.len() == pats.len()
+                    && p2.iter().all(|p| {
+                        matches!(p, IrPattern::Wildcard | IrPattern::None)
+                            || matches!(p, IrPattern::Constructor { args, .. } if args.is_empty())
+                    }) => {}
+            _ => return None,
+        }
+        Some(cond_n)
+    }
+
+    /// One tuple-refinement rewrite at `e` (the 2-arm conditional-column
+    /// shape): hoist each non-Var component once, bind the unconditional
+    /// components, and nest the conditional columns right-to-left. True =
+    /// rewritten in place.
+    fn rewrite_tuple_refinement(
+        e: &mut IrExpr,
+        next: &mut u32,
+        conditional: fn(&IrPattern) -> bool,
+    ) -> bool {
+            let IrExprKind::Match { subject, arms } = &e.kind else { return false };
+            let IrExprKind::Tuple { elements } = &subject.kind else { return false };
             if elements.len() < 2 || arms.len() != 2 || arms.iter().any(|a| a.guard.is_some()) {
-                return;
+                return false;
             }
-            let IrPattern::Tuple { elements: pats } = &arms[0].pattern else { return };
+            let IrPattern::Tuple { elements: pats } = &arms[0].pattern else { return false };
             if pats.len() != elements.len() {
-                return;
+                return false;
             }
-            let mut cond_n = 0usize;
-            for p in pats {
-                if conditional(p) {
-                    cond_n += 1;
-                } else if !matches!(p, IrPattern::Wildcard | IrPattern::Bind { .. }) {
-                    return;
-                }
-            }
-            if cond_n == 0 {
-                return;
-            }
-            match &arms[1].pattern {
-                IrPattern::Wildcard => {}
-                IrPattern::Tuple { elements: p2 }
-                    if p2.len() == pats.len()
-                        && p2.iter().all(|p| {
-                            matches!(p, IrPattern::Wildcard | IrPattern::None)
-                                || matches!(p, IrPattern::Constructor { args, .. } if args.is_empty())
-                        }) => {}
-                _ => return,
-            }
+            let Some(cond_n) = refinement_column_census(pats, &arms[1].pattern, conditional) else {
+                return false;
+            };
             let els = &arms[1].body;
             if cond_n > 1 && introduces_binder(els) {
-                return;
+                return false;
             }
             let span = e.span.clone();
-            // Hoist each non-Var component ONCE into a temp (a Var component is used direct).
-            let mut stmts: Vec<IrStmt> = Vec::new();
-            let mut refs: Vec<IrExpr> = Vec::new();
-            for c in elements {
-                if matches!(c.kind, IrExprKind::Var { .. }) {
-                    refs.push(c.clone());
-                } else {
-                    let t = VarId(self.next);
-                    self.next += 1;
-                    stmts.push(IrStmt {
-                        kind: IrStmtKind::Bind {
-                            var: t,
-                            ty: c.ty.clone(),
-                            value: c.clone(),
-                            mutability: almide_ir::Mutability::Let,
-                        },
-                        span: span.clone(),
-                    });
-                    refs.push(IrExpr {
-                        kind: IrExprKind::Var { id: t },
-                        ty: c.ty.clone(),
-                        span: span.clone(),
-                        def_id: None,
-                    });
-                }
-            }
+            let (stmts, refs) = hoist_tuple_components(elements, next, &span);
             // Innermost THEN: arm-1's body prefixed by its unconditional component binds.
             let mut binds: Vec<IrStmt> = Vec::new();
             for (i, p) in pats.iter().enumerate() {
@@ -341,32 +412,7 @@ pub fn desugar_tuple_variant_match(body: &IrExpr) -> Option<IrExpr> {
                     def_id: arms[0].body.def_id,
                 }
             };
-            // Nest the conditional components right-to-left (leftmost test outermost).
-            for (i, p) in pats.iter().enumerate().rev() {
-                if !conditional(p) {
-                    continue;
-                }
-                inner = IrExpr {
-                    kind: IrExprKind::Match {
-                        subject: Box::new(refs[i].clone()),
-                        arms: vec![
-                            almide_ir::IrMatchArm {
-                                pattern: p.clone(),
-                                guard: Option::None,
-                                body: inner,
-                            },
-                            almide_ir::IrMatchArm {
-                                pattern: IrPattern::Wildcard,
-                                guard: Option::None,
-                                body: els.clone(),
-                            },
-                        ],
-                    },
-                    ty: e.ty.clone(),
-                    span: span.clone(),
-                    def_id: e.def_id,
-                };
-            }
+            inner = nest_conditional_columns(NestCtx { e, span: span.clone(), pats, refs: &refs, els, conditional }, inner);
             *e = if stmts.is_empty() {
                 inner
             } else {
@@ -377,7 +423,19 @@ pub fn desugar_tuple_variant_match(body: &IrExpr) -> Option<IrExpr> {
                     def_id: e.def_id,
                 }
             };
-            self.changed = true;
+        true
+    }
+
+    struct V {
+        next: u32,
+        changed: bool,
+    }
+    impl IrMutVisitor for V {
+        fn visit_expr_mut(&mut self, e: &mut IrExpr) {
+            walk_expr_mut(self, e);
+            if rewrite_tuple_refinement(e, &mut self.next, conditional) {
+                self.changed = true;
+            }
         }
     }
     let mut v = V { next: max_var_id(body) + 1, changed: false };
@@ -386,415 +444,4 @@ pub fn desugar_tuple_variant_match(body: &IrExpr) -> Option<IrExpr> {
     v.changed.then_some(out)
 }
 
-/// N-ARM tuple-of-variants match — the Maranget-style column specialization the 2-arm
-/// [`desugar_tuple_variant_match`] (which runs FIRST in both chains) declines: 3+ arms, a
-/// binder-carrying fall-through arm (`(Leaf(a), Leaf(b)) => …, (l, r) => …` — the #610
-/// in-group refinement the deep variant regroup emits), and arbitrary ctor DEPTH in any
-/// component (`(Leaf(a), Node(Leaf(b), Leaf(c)))`). Recursively specialize the LEFTMOST
-/// conditional column into one single-subject match per ctor head: the head's payload
-/// fields bind to FRESH vars (new columns), a row whose column is Bind/Wildcard joins
-/// EVERY head's branch (its Bind substituted by the component ref — no duplicate binder),
-/// and the trivial-column rows form the `_` default — OMITTED when the heads cover the
-/// component's type exhaustively (a reachable-only-through-covered-heads default would
-/// embed a NON-exhaustive inner match and wall the whole fn). First-match order is
-/// preserved inside every branch; rows after the first all-trivial row prune. A body
-/// cloned into >1 branch must not introduce binders ([`introduces_binder`] — VarId
-/// uniqueness under duplication); Literal / record / list components decline (the literal
-/// tuple chain and the `[]`-column specializer own those). Runs in BOTH chains
-/// (desugar-before-both), so duplicated bodies count 1:1 in the caps `mir == ir` gate.
-pub fn desugar_tuple_variant_match_deep(
-    body: &IrExpr,
-    layouts: &crate::lower::VariantLayouts,
-) -> Option<IrExpr> {
-    use almide_ir::visit_mut::{walk_expr_mut, IrMutVisitor};
-    use almide_ir::{IrMatchArm, IrPattern};
-    use almide_lang::types::constructor::TypeConstructorId;
-    use almide_lang::types::Ty;
-
-    /// A dispatchable constructor head in a tuple column. (Guards/Literals/records never
-    /// reach here — `comp_ok` gates them out before compilation.)
-    #[derive(Clone, PartialEq, Eq)]
-    enum HKey {
-        User(String),
-        Some_,
-        None_,
-        Ok_,
-        Err_,
-    }
-    fn head_of(p: &IrPattern) -> Option<(HKey, Vec<IrPattern>)> {
-        match p {
-            IrPattern::Constructor { name, args } => {
-                Some((HKey::User(name.clone()), args.clone()))
-            }
-            IrPattern::Some { inner } => Some((HKey::Some_, vec![(**inner).clone()])),
-            IrPattern::None => Some((HKey::None_, vec![])),
-            IrPattern::Ok { inner } => Some((HKey::Ok_, vec![(**inner).clone()])),
-            IrPattern::Err { inner } => Some((HKey::Err_, vec![(**inner).clone()])),
-            _ => None,
-        }
-    }
-    fn trivial(p: &IrPattern) -> bool {
-        matches!(p, IrPattern::Wildcard | IrPattern::Bind { .. })
-    }
-    fn comp_ok(p: &IrPattern) -> bool {
-        trivial(p)
-            || head_of(p).is_some_and(|(_, args)| args.iter().all(comp_ok))
-    }
-    /// The declared payload types of `key` when the component has type `cty` — `None`
-    /// declines (unknown ctor, arity drift, a still-generic layout).
-    fn head_field_tys(
-        key: &HKey,
-        arity: usize,
-        cty: &Ty,
-        layouts: &crate::lower::VariantLayouts,
-    ) -> Option<Vec<Ty>> {
-        match key {
-            HKey::User(name) => {
-                let (tyname, layout, case) = layouts.lookup_ctor(name)?;
-                let _ = tyname;
-                if !layout.generics.is_empty() || case.fields.len() != arity {
-                    return None;
-                }
-                Some(case.fields.iter().map(|(_, t)| t.clone()).collect())
-            }
-            HKey::Some_ => match cty {
-                Ty::Applied(TypeConstructorId::Option, a) if a.len() == 1 => {
-                    Some(vec![a[0].clone()])
-                }
-                _ => None,
-            },
-            HKey::Ok_ => match cty {
-                Ty::Applied(TypeConstructorId::Result, a) if a.len() == 2 => {
-                    Some(vec![a[0].clone()])
-                }
-                _ => None,
-            },
-            HKey::Err_ => match cty {
-                Ty::Applied(TypeConstructorId::Result, a) if a.len() == 2 => {
-                    Some(vec![a[1].clone()])
-                }
-                _ => None,
-            },
-            HKey::None_ => Some(vec![]),
-        }
-    }
-    /// Do `keys` cover the component's type EXHAUSTIVELY (so the emitted match needs no
-    /// `_` arm)? Conservative: an unresolvable/generic layout answers `false` (the caller
-    /// then requires a real default or declines).
-    fn heads_cover(keys: &[HKey], layouts: &crate::lower::VariantLayouts) -> bool {
-        if keys.iter().all(|k| matches!(k, HKey::Some_ | HKey::None_)) {
-            return keys.contains(&HKey::Some_) && keys.contains(&HKey::None_);
-        }
-        if keys.iter().all(|k| matches!(k, HKey::Ok_ | HKey::Err_)) {
-            return keys.contains(&HKey::Ok_) && keys.contains(&HKey::Err_);
-        }
-        if !keys.iter().all(|k| matches!(k, HKey::User(_))) {
-            return false;
-        }
-        let HKey::User(first) = &keys[0] else { return false };
-        let Some(tyname) = layouts.ctor_to_type.get(first) else { return false };
-        let Some(layout) = layouts.by_type.get(tyname) else { return false };
-        !layout.cases.is_empty()
-            && layout.cases.iter().all(|c| {
-                keys.iter().any(|k| matches!(k, HKey::User(n) if n == c.ctor.as_str()))
-            })
-    }
-
-    struct Row {
-        pats: Vec<IrPattern>,
-        body: IrExpr,
-        idx: usize,
-    }
-    /// The recursive column compiler. `refs[i]` is the (Var) expression re-reading column
-    /// `i`; `tmpl` supplies the result ty/span/def_id; `emitted[idx]` counts how many
-    /// branches cloned original arm `idx`'s body (the duplication gate reads it after).
-    fn compile(
-        refs: &[IrExpr],
-        mut rows: Vec<Row>,
-        tmpl: &IrExpr,
-        next: &mut u32,
-        layouts: &crate::lower::VariantLayouts,
-        emitted: &mut Vec<usize>,
-    ) -> Option<IrExpr> {
-        // First-match pruning: rows after the first all-trivial (always-matching) row are dead.
-        if let Some(k) = rows.iter().position(|r| r.pats.iter().all(trivial)) {
-            rows.truncate(k + 1);
-        }
-        let first_all_trivial = rows.first()?.pats.iter().all(trivial);
-        if first_all_trivial {
-            return Some(trivial_row_body(&rows[0], refs, emitted));
-        }
-        let j = (0..refs.len()).find(|&c| rows.iter().any(|r| !trivial(&r.pats[c])))?;
-        let rows = substitute_dispatch_column_binds(rows, refs, j);
-        let keys = ordered_ctor_heads(&rows, j)?;
-        let mut arms: Vec<IrMatchArm> = Vec::new();
-        for (key, arity) in &keys {
-            let ftys = head_field_tys(key, *arity, &refs[j].ty, layouts)?;
-            let fresh: Vec<(VarId, Ty)> = ftys
-                .iter()
-                .map(|t| {
-                    let v = VarId(*next);
-                    *next += 1;
-                    (v, t.clone())
-                })
-                .collect();
-            let nrefs = refs_with_head_fields(refs, j, &fresh, tmpl);
-            let nrows = specialize_rows_for_head(&rows, j, key, *arity, nrefs.len());
-            let branch = compile(&nrefs, nrows, tmpl, next, layouts, emitted)?;
-            let pat_args: Vec<IrPattern> = fresh
-                .iter()
-                .map(|(v, t)| IrPattern::Bind { var: *v, ty: t.clone() })
-                .collect();
-            arms.push(IrMatchArm { pattern: head_pattern(key, pat_args), guard: None, body: branch });
-        }
-        let head_keys: Vec<HKey> = keys.iter().map(|(k, _)| k.clone()).collect();
-        if !heads_cover(&head_keys, layouts) {
-            let drows = rows_without_dispatch_column(&rows, j);
-            if drows.is_empty() {
-                // Frontend exhaustiveness says this path is unreachable, but emitting a
-                // non-exhaustive inner match would wall — decline instead.
-                return None;
-            }
-            let mut nrefs = refs.to_vec();
-            nrefs.remove(j);
-            let dbody = compile(&nrefs, drows, tmpl, next, layouts, emitted)?;
-            arms.push(IrMatchArm { pattern: IrPattern::Wildcard, guard: None, body: dbody });
-        }
-        Some(IrExpr {
-            kind: IrExprKind::Match { subject: Box::new(refs[j].clone()), arms },
-            ty: tmpl.ty.clone(),
-            span: tmpl.span.clone(),
-            def_id: tmpl.def_id,
-        })
-    }
-
-    /// The LEAF of [`compile`]: an all-trivial row matches unconditionally, so its body IS
-    /// the branch — with each `Bind` pattern's var substituted by the column ref it names,
-    /// and the row's duplication count bumped (the gate reads it after). Extracted verbatim
-    /// (codopsy round-3 sweep, #852).
-    fn trivial_row_body(r: &Row, refs: &[IrExpr], emitted: &mut [usize]) -> IrExpr {
-        let mut b = r.body.clone();
-        for (i, p) in r.pats.iter().enumerate() {
-            if let IrPattern::Bind { var, .. } = p {
-                b = almide_ir::substitute_var_in_expr(&b, *var, &refs[i]);
-            }
-        }
-        emitted[r.idx] += 1;
-        b
-    }
-
-    /// A `Bind` in the DISPATCH column names the WHOLE component: substitute the component
-    /// ref now (once, before the row joins multiple branches) and dispatch on `_`. Extracted
-    /// verbatim from [`compile`] (codopsy round-3 sweep, #852).
-    fn substitute_dispatch_column_binds(rows: Vec<Row>, refs: &[IrExpr], j: usize) -> Vec<Row> {
-        rows.into_iter()
-            .map(|mut r| {
-                if let IrPattern::Bind { var, .. } = &r.pats[j] {
-                    r.body = almide_ir::substitute_var_in_expr(&r.body, *var, &refs[j]);
-                    r.pats[j] = IrPattern::Wildcard;
-                }
-                r
-            })
-            .collect()
-    }
-
-    /// The dispatch column's ctor heads in FIRST-OCCURRENCE order, with each head's arity.
-    /// `None` declines the whole compile: a same-head arity drift cannot be dispatched.
-    /// Extracted verbatim from [`compile`] (codopsy round-3 sweep, #852).
-    fn ordered_ctor_heads(rows: &[Row], j: usize) -> Option<Vec<(HKey, usize)>> {
-        let mut keys: Vec<(HKey, usize)> = Vec::new();
-        for r in rows {
-            if let Some((k, args)) = head_of(&r.pats[j]) {
-                match keys.iter().find(|(k2, _)| *k2 == k) {
-                    Some((_, a)) if *a != args.len() => return None,
-                    Some(_) => {}
-                    None => keys.push((k, args.len())),
-                }
-            }
-        }
-        Some(keys)
-    }
-
-    /// The column refs ONE head's branch sees: the dispatch column replaced in place by the
-    /// head's freshly-bound field refs. Extracted verbatim from [`compile`] (codopsy round-3
-    /// sweep, #852).
-    fn refs_with_head_fields(
-        refs: &[IrExpr],
-        j: usize,
-        fresh: &[(VarId, Ty)],
-        tmpl: &IrExpr,
-    ) -> Vec<IrExpr> {
-        let mut nrefs: Vec<IrExpr> = Vec::with_capacity(refs.len() - 1 + fresh.len());
-        nrefs.extend_from_slice(&refs[..j]);
-        for (v, t) in fresh {
-            nrefs.push(IrExpr {
-                kind: IrExprKind::Var { id: *v },
-                ty: t.clone(),
-                span: tmpl.span.clone(),
-                def_id: None,
-            });
-        }
-        nrefs.extend_from_slice(&refs[j + 1..]);
-        nrefs
-    }
-
-    /// The rows ONE head's branch sees: a row whose head IS this key contributes its head
-    /// args in the dispatch column's place; a HEADLESS (trivial) row contributes wildcards
-    /// there (it still matches); a row with a DIFFERENT head is dropped. Extracted verbatim
-    /// from [`compile`] (codopsy round-3 sweep, #852).
-    fn specialize_rows_for_head(
-        rows: &[Row],
-        j: usize,
-        key: &HKey,
-        arity: usize,
-        width: usize,
-    ) -> Vec<Row> {
-        let mut nrows: Vec<Row> = Vec::new();
-        for r in rows {
-            match head_of(&r.pats[j]) {
-                Some((k, args)) if k == *key => {
-                    let mut np = Vec::with_capacity(width);
-                    np.extend_from_slice(&r.pats[..j]);
-                    np.extend(args);
-                    np.extend_from_slice(&r.pats[j + 1..]);
-                    nrows.push(Row { pats: np, body: r.body.clone(), idx: r.idx });
-                }
-                Some(_) => {}
-                None => {
-                    let mut np = Vec::with_capacity(width);
-                    np.extend_from_slice(&r.pats[..j]);
-                    np.extend(std::iter::repeat(IrPattern::Wildcard).take(arity));
-                    np.extend_from_slice(&r.pats[j + 1..]);
-                    nrows.push(Row { pats: np, body: r.body.clone(), idx: r.idx });
-                }
-            }
-        }
-        nrows
-    }
-
-    /// One head key as the arm PATTERN binding its fresh field vars. Extracted verbatim
-    /// from [`compile`] (codopsy round-3 sweep, #852).
-    fn head_pattern(key: &HKey, mut pat_args: Vec<IrPattern>) -> IrPattern {
-        match key {
-            HKey::User(name) => IrPattern::Constructor { name: name.clone(), args: pat_args },
-            HKey::Some_ => IrPattern::Some { inner: Box::new(pat_args.remove(0)) },
-            HKey::None_ => IrPattern::None,
-            HKey::Ok_ => IrPattern::Ok { inner: Box::new(pat_args.remove(0)) },
-            HKey::Err_ => IrPattern::Err { inner: Box::new(pat_args.remove(0)) },
-        }
-    }
-
-    /// The DEFAULT arm's rows: the headless rows only, with the dispatch column removed
-    /// (nothing tested it). Extracted verbatim from [`compile`] (codopsy round-3 sweep,
-    /// #852).
-    fn rows_without_dispatch_column(rows: &[Row], j: usize) -> Vec<Row> {
-        let mut drows: Vec<Row> = Vec::new();
-        for r in rows {
-            if head_of(&r.pats[j]).is_none() {
-                let mut np = r.pats.clone();
-                np.remove(j);
-                drows.push(Row { pats: np, body: r.body.clone(), idx: r.idx });
-            }
-        }
-        drows
-    }
-
-
-    struct V<'a> {
-        next: u32,
-        layouts: &'a crate::lower::VariantLayouts,
-        changed: bool,
-    }
-    impl IrMutVisitor for V<'_> {
-        fn visit_expr_mut(&mut self, e: &mut IrExpr) {
-            walk_expr_mut(self, e);
-            let IrExprKind::Match { subject, arms } = &e.kind else { return };
-            let IrExprKind::Tuple { elements } = &subject.kind else { return };
-            let n = elements.len();
-            if n < 2 || arms.is_empty() || arms.iter().any(|a| a.guard.is_some()) {
-                return;
-            }
-            // Normalize arms to pattern ROWS: a Tuple pattern of matching width, or a
-            // trailing top-level `_` (an all-wildcard row). Anything else declines.
-            let mut rows: Vec<Row> = Vec::with_capacity(arms.len());
-            let mut any_cond = false;
-            for (idx, a) in arms.iter().enumerate() {
-                let pats: Vec<IrPattern> = match &a.pattern {
-                    IrPattern::Tuple { elements: ps } if ps.len() == n => ps.clone(),
-                    IrPattern::Wildcard => vec![IrPattern::Wildcard; n],
-                    _ => return,
-                };
-                if !pats.iter().all(comp_ok) {
-                    return;
-                }
-                if pats.iter().any(|p| !trivial(p)) {
-                    any_cond = true;
-                }
-                rows.push(Row { pats, body: a.body.clone(), idx });
-            }
-            if !any_cond {
-                return;
-            }
-            // Hoist each non-Var component ONCE into a temp (a Var component reads direct).
-            let span = e.span.clone();
-            let mut stmts: Vec<IrStmt> = Vec::new();
-            let mut refs: Vec<IrExpr> = Vec::new();
-            for c in elements {
-                if matches!(c.kind, IrExprKind::Var { .. }) {
-                    refs.push(c.clone());
-                } else {
-                    let t = VarId(self.next);
-                    self.next += 1;
-                    stmts.push(IrStmt {
-                        kind: IrStmtKind::Bind {
-                            var: t,
-                            ty: c.ty.clone(),
-                            value: c.clone(),
-                            mutability: almide_ir::Mutability::Let,
-                        },
-                        span: span.clone(),
-                    });
-                    refs.push(IrExpr {
-                        kind: IrExprKind::Var { id: t },
-                        ty: c.ty.clone(),
-                        span: span.clone(),
-                        def_id: None,
-                    });
-                }
-            }
-            let mut emitted = vec![0usize; arms.len()];
-            let mut next = self.next;
-            let Some(compiled) =
-                compile(&refs, rows, e, &mut next, self.layouts, &mut emitted)
-            else {
-                return;
-            };
-            // Duplication gates: a body cloned into >1 branch must be binder-free, and the
-            // whole tree must stay small (the same blow-up discipline as heap-branches).
-            for (idx, count) in emitted.iter().enumerate() {
-                if *count > 1 && introduces_binder(&arms[idx].body) {
-                    return;
-                }
-            }
-            if count_expr_nodes(&compiled) > 50_000 {
-                return;
-            }
-            self.next = next;
-            *e = if stmts.is_empty() {
-                compiled
-            } else {
-                IrExpr {
-                    kind: IrExprKind::Block { stmts, expr: Some(Box::new(compiled)) },
-                    ty: e.ty.clone(),
-                    span,
-                    def_id: e.def_id,
-                }
-            };
-            self.changed = true;
-        }
-    }
-    let mut v = V { next: max_var_id(body) + 1, layouts, changed: false };
-    let mut out = body.clone();
-    v.visit_expr_mut(&mut out);
-    v.changed.then_some(out)
-}
+include!("desugar_match_deep.rs");
